requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
av==14.2.0
bcrypt==4.3.0
blake3==1.0.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.2
compressed-tensors==0.9.2
ConfigArgParse==1.7
contourpy==1.3.1
cryptography==44.0.2
cupy-cuda12x==13.4.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
decorator==5.1.1
defusedxml==0.7.1
depyf==0.18.0
dill==0.3.9
diskcache==5.6.3
distro==1.7.0
dnspython==2.7.0
echo==0.9.0
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.17.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gguf==0.10.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
interegular==0.3.3
ipdb==0.13.13
ipython==8.32.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kmeans-pytorch==0.3
lark==1.2.2
llguidance==0.7.11
llvmlite==0.43.0
lm-format-enforcer==0.10.11
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.4
mpi4py==3.1.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.2.0
narwhals==1.26.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
onnx==1.16.1
onnxruntime-gpu==1.18.0
open3d==0.19.0
openai==1.69.0
opencv-python==4.11.0.86
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parse==1.20.2
parso==0.8.4
partial-json-parser==0.2.1.1.post5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycairo==1.20.1
pycocotools==2.0.8
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyGObject==3.42.1
PyNaCl==1.5.0
PyOpenGL==3.1.9
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyquaternion==0.9.9
pyrealsense2==2.55.1.6486
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
qwen-vl-utils==0.0.10
ray==2.44.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rich-toolkit==0.14.0
rpds-py==0.22.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
scp==0.15.0
seaborn==0.13.2
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
-e git+ssh://git@github.com/HKCLR2021/Blueberry-Picking-UR.git@fb95355a00beb7a618009062fe3934801276b9a4#egg=segmentation&subdirectory=segmentation
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
stack-data==0.6.3
starlette==0.46.1
supervision==0.25.1
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
torchaudio==2.6.0+cu118
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/qubvel/transformers@a0ed339ce114cd3e48e839604248352e0b29065f
transforms3d==0.4.2
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.37
ultralytics-thop==2.0.14
-e git+https://gitlab.com/sdurobotics/ur_rtde.git@1b059b1e46e8d46c4ed6bbd85b7dd179f81bdd62#egg=ur_rtde
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
visual-kinematics==0.2.1
vllm==0.8.2
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
xformers==0.0.29.post2
xgrammar==0.1.16
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
